fix(web): gate start-agent hint on composer availability
