+++ apps/web/e2e/tests/settings/prevent-auto-start-on-open.spec.ts
@@ -167,12 +167,15 @@ test.describe("Prevent auto-start on open", () => {
       await expect(testPage.getByText("Resumed agent", { exact: false })).toHaveCount(0);
 
       // Sending a message is the explicit start: the agent resumes and
-      // keeps working.
+      // keeps working. Index 1: the first "simple mock response" from the
+      // pre-restart turn is still in the transcript, so the post-send reply
+      // must be the SECOND match — asserting index 0 would pass even if the
+      // resumed launch dropped the prompt.
       await session.sendMessage("/e2e:simple-message");
       await expect(testPage.getByText("Resumed agent", { exact: false })).toBeVisible({
         timeout: 30_000,
       });
-      await session.expectChatResponseVisible("simple mock response", 0, { timeout: 30_000 });
+      await session.expectChatResponseVisible("simple mock response", 1, { timeout: 30_000 });
     } finally {
       await apiClient.saveUserSettings({ prevent_auto_start_agent_on_open: false });
     }
